TECH-2322 expose es6 modules: collapse unnecessary DataTableFilter function
